Move app test data from main.js to data.js; Move import util functions from main.js to data.js; Export interface function from data.js to main.js; import interface function to main.js;
